--- src/mock/offers.js
@@ -1,134 +1,0 @@
-export const OFFERS = [
-  {
-    type: 'taxi',
-    offers: [
-      {
-        id: 1,
-        title: 'Switch to business',
-        price: 30
-      },
-      {
-        id: 2,
-        title: 'Smoke inside',
-        price: 10
-      },
-      {
-        id: 3,
-        title: 'Play my music',
-        price: 100
-      }
-    ]
-  },
-  {
-    type: 'bus',
-    offers: null
-  },
-  {
-    type: 'train',
-    offers: null
-  },
-  {
-    type: 'ship',
-    offers: null
-  },
-  {
-    type: 'drive',
-    offers: [
-      {
-        id: 1,
-        title: 'Full of gas',
-        price: 70
-      },
-      {
-        id: 2,
-        title: 'Drop the car everywhere',
-        price: 35
-      },
-      {
-        id: 3,
-        title: 'Unlimited insurance',
-        price: 90
-      }
-    ]
-  },
-  {
-    type: 'flight',
-    offers: [
-      {
-        id: 1,
-        title: 'Choose seats',
-        price: 5
-      },
-      {
-        id: 2,
-        title: 'Add luggage',
-        price: 50
-      },
-      {
-        id: 3,
-        title: 'Switch to comfort',
-        price: 80
-      }
-    ]
-  },
-  {
-    type: 'check-in',
-    offers: [
-      {
-        id: 1,
-        title: 'Pick up my stuff',
-        price: 15
-      },
-      {
-        id: 2,
-        title: 'Bar',
-        price: 50
-      },
-      {
-        id: 3,
-        title: 'Safe deposit',
-        price: 30
-      }
-    ]
-  },
-  {
-    type: 'sightseeing',
-    offers: [
-      {
-        id: 1,
-        title: 'Hire a guide',
-        price: 100
-      },
-      {
-        id: 2,
-        title: 'Get souvenir',
-        price: 20
-      },
-      {
-        id: 3,
-        title: 'No advertisement',
-        price: 15
-      }
-    ]
-  },
-  {
-    type: 'restaurant',
-    offers: [
-      {
-        id: 1,
-        title: 'Smoke hookah',
-        price: 35
-      },
-      {
-        id: 2,
-        title: 'Tips',
-        price: 5
-      },
-      {
-        id: 3,
-        title: 'Personal room',
-        price: 90
-      }
-    ]
-  },
-];
